requirements_full.txt
asgiref==3.8.1
attrs==24.2.0
boto3==1.34.72
django-storages==1.14.2
dj-database-url==2.1.0
Django==5.1.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
# Using binary version that doesn't require postgresql-devel
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.9.0 